% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkData}
\alias{checkData}
\title{Check the iris data.}
\usage{
checkData(iris.data)
}
\arguments{
\item{iris.data}{The \code{\link{iris.data}}}
}
\value{
Cleaned \code{\link{iris.data}}
}
\description{
Performs any necessary data cleaning.
}
\examples{
iris.data <- checkData(iris.data)
}
\author{
phil
}

