% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier}
\alias{classifier}
\title{Majority voting.}
\usage{
classifier(cls, distance)
}
\arguments{
\item{cls}{A list of neighbour classes found with \code{\link{knn}}.}

\item{distance}{An equal length list of neighbour distances.}
}
\value{
The majority prediction along with classification confidence.
}
\description{
Predict a class given a list of neighbours obtained from k-nn using majority
voting.
}
\details{
Todo: Distance weighted majority voting/distance kernel.
}
\examples{
# form the query instance.
query <- list(
    sepal.length=5.84,
    sepal.width=3.05,
    petal.length=3.76,
    petal.width=1.20)

# get the 10-nearest neighbours.
top.10 <- knn(query, iris.data, 10)

# classify the instance.
prediction <- classifier(top.10$species, top.10$distance)

# print the result.
print(paste("prediction =", prediction$pred,
            "confidence =", prediction$conf))
}
\author{
phil
}

