% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance metric.}
\usage{
euclidean(v1, v2)
}
\arguments{
\item{v1}{Source vector.}

\item{v2}{Target vector.}
}
\value{
Euclidean distance.
}
\description{
Euclidean distance between two vectors.
}
\details{
Note: There is also a \code{\link{dist}} function in R with the options for
euclidian, maximum, manhatten, canberra, binary and minkowski metrics.
}
\examples{
# (3,4) with (9,12)
euclidean(c(3, 4), c(9, 12))

# (3,4) with (44,66)
euclidean(c(3, 4), c(44, 66))

# (3,4) with (9,12) and (44,66)
m <- matrix(c(9, 12, 44, 66), ncol=2, byrow=TRUE)
apply(m, 1, euclidean, c(3, 4))
}
\author{
phil
}

