% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazy-iris.R
\docType{data}
\name{iris.data}
\alias{iris.data}
\title{Iris data.}
\format{A data.frame.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Iris}
}
\usage{
data(iris.data)
}
\description{
The Iris dataset.
}
\details{
Descripton from UCI Machine learning repository: This is perhaps the best
known database to be found in the pattern recognition literature. Fisher's
paper is a classic in the field and is referenced frequently to this day.
(See Duda & Hart, for example.) The data set contains 3 classes of 50
instances each, where each class refers to a type of iris plant. One class
is linearly separable from the other 2; the latter are NOT linearly
separable from each other.
}
\author{
phil
}
\references{
Lichman, M. (2013). UCI Machine Learning Repository
    \url{https://archive.ics.uci.edu/ml}. Irvine, CA: University of
    California, School of Information and Computer Science.
}
\keyword{datasets}

