% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K-Nearest-Neighbors.}
\usage{
knn(query, feature.space, k = 1, d.fun = euclidean)
}
\arguments{
\item{query}{A list of labeled parameters to query for.}

\item{feature.space}{The feature space to query against.}

\item{k}{The \emph{k} nearest neighbours to return. Default: 1.}

\item{d.fun}{The distance function to use. Default: \code{\link{euclidean}}.}
}
\value{
The top \emph{k} nearest neighours in the feature space with respect
    to the query.
}
\description{
Implementation of the
\href{https://en.wikipedia.org/wiki/K-nearest_neighbors_algorithm}{k-NN}
method.
}
\examples{
# form the query instance.
query <- list(
    sepal.length=5.84,
    sepal.width=3.05,
    petal.length=3.76,
    petal.width=1.20)

# obtain the nearest-neighbour
top.1 <- knn(query, iris.data, 1)
print(top.1, row.names=FALSE)
}
\author{
phil
}

