% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazy-iris.R
\docType{package}
\name{lazyIris-package}
\alias{lazyIris}
\alias{lazyIris-package}
\title{lazyIris.}
\description{
Lazy learning on the iris dataset.
}
\details{
k-nearest neighbour.
}
\section{Functionality}{


\itemize{
  \item Read in the data.
  \item Accesses the quality of the data.
  \item Returns the ten most similar data points in the existing iris data
given inputted arguments.
  \item Visualises the result.
}
}
\author{
phil \email{phil@parasec.net}
}

