% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadData}
\alias{loadData}
\title{Load data.}
\usage{
loadData()
}
\value{
Iris data as documented in \code{\link{iris.data}}
}
\description{
Loads the iris csv data from the inst/extdata directory.
}
\examples{
iris.data <- loadData()
}
\author{
phil
}

